bernoulli likelihood

+++ ModulatedGPs/models.py
@@ -78,7 +78,7 @@ class SMGP(SGP):
 
     def E_log_p_Y(self, Xt, Y, W_SND):
         Fmean, Fvar = self._build_predict(Xt, full_cov=False)
-        var_exp = self.likelihood.variational_expectations([], Fmean, Fvar, Y)
+        var_exp = self.likelihood.variational_expectations([], Fmean, Fvar, tf.cast(Y, dtype=tf.float64))
         var_exp *= tf.cast(W_SND, dtype=float_type)
         return tf.reduce_logsumexp(tf.reduce_sum(var_exp, 2), 0) - np.log(self.num_samples)
